# api/bundled/openapi.yaml
openapi: 3.0.3
info:
  title: Moneo Backend API (as-is)
  version: 0.1.0
  license:
    name: Proprietary
    url: https://example.com/license
  description: |
    Contract documents current behavior of auth, catalogs and transactions endpoints as implemented.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
  /api/v1/auth/register:
    post:
      summary: POST operation
      tags:
        - Auth
      operationId: registerAuth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/AuthError'
        '409':
          $ref: '#/components/responses/AuthError'
        '500':
          $ref: '#/components/responses/AuthError'
  /api/v1/auth/login:
    post:
      summary: POST operation
      tags:
        - Auth
      operationId: loginAuth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/AuthError'
        '401':
          $ref: '#/components/responses/AuthError'
        '500':
          $ref: '#/components/responses/AuthError'
  /api/v1/auth/refresh:
    post:
      summary: POST operation
      tags:
        - Auth
      operationId: refreshAuth
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/AuthError'
        '401':
          $ref: '#/components/responses/AuthError'
        '500':
          $ref: '#/components/responses/AuthError'
  /api/v1/auth/logout:
    post:
      summary: POST operation
      tags:
        - Auth
      operationId: logoutAuth
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthOKResponse'
        '400':
          $ref: '#/components/responses/AuthError'
        '500':
          $ref: '#/components/responses/AuthError'
  /api/v1/auth/logout-all:
    post:
      summary: POST operation
      tags:
        - Auth
      operationId: logoutAllAuth
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthOKResponse'
        '401':
          $ref: '#/components/responses/AuthError'
        '500':
          $ref: '#/components/responses/AuthError'
...
  /api/v1/auth/forgot-password:
    post:
      summary: POST operation
      tags:
        - Auth
      operationId: forgotPasswordAuth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthForgotPasswordRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthOKResponse'
        '400':
          $ref: '#/components/responses/AuthError'
        '500':
          $ref: '#/components/responses/AuthError'
  /api/v1/auth/reset-password:
    post:
      summary: POST operation
      tags:
        - Auth
      operationId: resetPasswordAuth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthResetPasswordRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthOKResponse'
        '400':
          $ref: '#/components/responses/AuthError'
        '401':
          $ref: '#/components/responses/AuthError'
        '500':
          $ref: '#/components/responses/AuthError'
...
  /api/v1/auth/verify-email:
    post:
      summary: POST operation
      tags:
        - Auth
      operationId: verifyEmailAuth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthVerifyEmailRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthOKResponse'
        '400':
          $ref: '#/components/responses/AuthError'
        '401':
          $ref: '#/components/responses/AuthError'
        '500':
          $ref: '#/components/responses/AuthError'
...
components:
  schemas:
    AuthRegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        password_confirm:
          type: string
    AuthResponse:
      type: object
      required:
        - access_token
        - expires_in
      properties:
        access_token:
          type: string
        expires_in:
          type: integer
    AuthErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    AuthLoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    AuthOKResponse:
      type: object
      required:
        - ok
      properties:
        ok:
          type: boolean
...
    AuthForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
    AuthResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
        password_confirm:
          type: string
    AuthVerifyEmailRequest:
      type: object
      properties:
        token:
          type: string
...
  responses:
    AuthError:
      description: Auth flow error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthErrorResponse'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT